moved the methods to the function.c file
